backend/requirements.txt
Django>=4.2
djangorestframework
requests
python-decouple